sketch, need to figure out 3d indexing into linear array

+++ src/distancefield.jl
@@ -10,15 +10,17 @@ function SignedDistanceField(primitive::AbstractPrimitive{3,T},
     ny = ceil(Int, y_rng/resolution)
     nz = ceil(Int, z_rng/resolution)
 
-    vol = Array{Float64}(undef,nx+1, ny+1, nz+1)
+    vol = Array{Float32}(undef,nx+1, ny+1, nz+1)
 
     b_max = SVector(x_min + resolution*nx, y_min + resolution*ny, z_min + resolution*nz)
 
     # re-adjust bounding box
-    o = SVector{3,Float64}(origin(bounds)...)
+    o = SVector{3,Float32}(origin(bounds)...)
     w = b_max-o
     bounds = HyperRectangle(o..., w...)
 
+    # TODO this seem like the dimensions will be off a little since
+    # e resize the bounds and do not update the resolution
     for i = 0:nx, j = 0:ny, k = 0:nz
         x = x_min + resolution*i
         y = y_min + resolution*j
@@ -26,5 +28,5 @@ function SignedDistanceField(primitive::AbstractPrimitive{3,T},
         @inbounds vol[i+1,j+1,k+1] = FRep(primitive,x,y,z)
     end
 
-    SignedDistanceField{3,Float64,Float64}(bounds, vol)
+    SignedDistanceField{3,Float32,Float32}(bounds, vol)
 end
